Move and update `DateFieldtype` rule class to `Statamic\Rules` w/ Laravel conventions.
