+++ src/Fieldtypes/Date.php
@@ -4,16 +4,16 @@ namespace Statamic\Fieldtypes;
 
 use Carbon\Exceptions\InvalidFormatException;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Validator;
 use InvalidArgumentException;
-use Statamic\Exceptions\ValidationException;
 use Statamic\Facades\GraphQL;
 use Statamic\Fields\Fieldtype;
 use Statamic\GraphQL\Fields\DateField;
 use Statamic\GraphQL\Types\DateRangeType;
 use Statamic\Query\Scopes\Filters\Fields\Date as DateFilter;
+use Statamic\Rules\DateFieldtype as ValidationRule;
 use Statamic\Statamic;
 use Statamic\Support\DateFormat;
-use Statamic\Validation\DateFieldtype as ValidationRule;
 
 class Date extends Fieldtype
 {
@@ -366,11 +366,10 @@ class Date extends Fieldtype
 
     public function preProcessValidatable($value)
     {
-        if ($error = (new ValidationRule($this))($value)) {
-            throw ValidationException::withMessages([
-                $this->field->handle() => $error,
-            ]);
-        }
+        Validator::make(
+            [$this->field->handle() => $value],
+            [$this->field->handle() => [new ValidationRule($this)]],
+        )->validate();
 
         if ($value === null) {
             return null;
